+++ projects/common/src/index.ts
@@ -32,7 +32,6 @@ export enum ArticleType {
     Letter = 'letter',
     Recipe = 'recipe',
     Gallery = 'gallery',
-    Picture = 'picture',
     Feature = 'feature',
     Immersive = 'immersive',
 }
